Rename currentIntersectionAtoms to selectionAtoms and move relevant selectors there
